third_party/WebKit/LayoutTests/html/dialog/inert-node-is-uneditable-expected.txt

Test that inert nodes cannot be edited. The test passes if the only text you can edit is in the dialog.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS notEditable.textContent is oldValue
PASS editable.textContent is not oldValue
PASS successfullyParsed is true

TEST COMPLETE
